+++ data/src/main/java/com/example/data/source/remote/ApiRemoteSource.kt
@@ -13,11 +13,11 @@ class ApiRemoteSource(
 
     suspend fun getPublicationById(id: Int): PublicationModel = api.getPublicationById(id)
 
-    suspend fun getListAllDraftPublication(cookie: String, channelId: Int?): List<PublicationModel> = api.getListAllDraftPublication(cookie, channelId)
+    suspend fun getListAllDraftPublication(cookie: String, channelId: Long?): List<PublicationModel> = api.getListAllDraftPublication(cookie, channelId)
 
-    suspend fun getListAllSendPublication(cookie: String, channelId: Int?): List<PublicationModel> = api.getListAllSendPublication(cookie, channelId)
+    suspend fun getListAllSendPublication(cookie: String, channelId: Long?): List<PublicationModel> = api.getListAllSendPublication(cookie, channelId)
 
-    suspend fun getListAllSchedulePublication(cookie: String, channelId: Int?): List<PublicationModel> = api.getListAllSchedulePublication(cookie, channelId)
+    suspend fun getListAllSchedulePublication(cookie: String, channelId: Long?): List<PublicationModel> = api.getListAllSchedulePublication(cookie, channelId)
 
     suspend fun getListChannels(cookie: String): List<ChannelModel> = api.getListChannels(cookie)
 
@@ -27,7 +27,7 @@ class ApiRemoteSource(
         api.sendMobileToken(token, cookie)
     }
 
-    suspend fun updatePost(cookie: String, id: Int, channelId: Int, name: String, text: String, comment: String, scheduledAt: String?) {
+    suspend fun updatePost(cookie: String, id: Int, channelId: Long, name: String, text: String, comment: String, scheduledAt: String?) {
         api.updatePost(cookie, id, channelId, name, text, comment, scheduledAt)
     }
